Add battery monitoring with BLE service and ADC calibration
- Add BatteryMonitor class with hardware-specific ADC calibration
  (curve fitting for ESP32-C3, line fitting for ESP32)
- Add BatteryServiceProfile implementing standard BLE Battery Service (0x180F)
- Add Kconfig options for battery GPIO, voltage divider, and thresholds
- Align all Docker configs to ESP-IDF v6.0-beta1
- Fix format specifiers for RISC-V cross-compilation
- Fix lambda captures for GCC 15 compatibility

+++ components/system_control/src/button_feedback.cpp
@@ -62,8 +62,8 @@ void ButtonFeedbackController::showFeedback(FeedbackType type) {
         feedback_duration_ms_ = duration_ms;
         feedback_active_ = true;
         
-        ESP_LOGD(TAG, "🎨 Button feedback sent to LED task: %s for %u ms", 
-                 getFeedbackTypeName(type), feedback_duration_ms_);
+        ESP_LOGD(TAG, "🎨 Button feedback sent to LED task: %s for %lu ms",
+                 getFeedbackTypeName(type), (unsigned long)feedback_duration_ms_);
     } else {
         ESP_LOGW(TAG, "Failed to send button feedback to LED task: %s", esp_err_to_name(result));
